Updated facrtory design
