+++ development/factory_design.cpp
@@ -1,80 +1,27 @@
-#include <iostream>
-#include <memory>
+#include "factory_design.hpp"
 
-class Product {
-public:
-    virtual void show() = 0;
-};
+// Implementation of DAQTransferA's show method
+void DAQTransferA::show() {
+    std::cout << "This is Product A\n";
+}
 
-class ProductA : public Product {
-public:
-    void show() override {
-        std::cout << "This is Product A\n";
-    }
-};
+// Implementation of DAQTransferB's show method
+void DAQTransferB::show() {
+    std::cout << "This is Product B\n";
+}
 
-class ProductB : public Product {
-public:
-    void show() override {
-        std::cout << "This is Product B\n";
-    }
-};
+// Implementation of DAQTransferFactory's getInstance method
+DAQTransferFactory& DAQTransferFactory::getInstance() {
+    static DAQTransferFactory instance;
+    return instance;
+}
 
-class ProductFactory {
-public:
-    static std::unique_ptr<Product> createProduct(const std::string& productType) {
-        if (productType == "A")
-            return std::unique_ptr<Product>(new ProductA());
-        else if (productType == "B")
-            return std::unique_ptr<Product>(new ProductB());
-        else
-            return nullptr;
-    }
-};
-
-
-class IDAQTransfer {
-public:
-    virtual void show() = 0;
-};
-
-class DAQTransferA : public IDAQTransfer {
-public:
-    void show() override {
-        std::cout << "This is Channel A\n";
-    }
-};
-
-class DAQTransferB : public IDAQTransfer {
-public:
-    void show() override {
-        std::cout << "This is Channel B\n";
-    }
-};
-
-class DAQTransferFactory {
-public:
-    //* Singleton pattern instance
-    static DAQTransferFactory& getInstance() {
-        static DAQTransferFactory instance;
-        return instance;
-    }
-
-    static std::unique_ptr<IDAQTransfer> createChannel(const std::string& channelType) {
-        if (channelType == "A")
-            return std::unique_ptr<IDAQTransfer>(new DAQTransferA());
-        else if (channelType == "B")
-            return std::unique_ptr<IDAQTransfer>(new DAQTransferB());
-        else
-            return nullptr;
-    }
-
-private:
-    //* Singleton pattern
-    // Private constructor to prevent instantiation
-    DAQTransferFactory() {}
-    // Delete copy constructor and assignment operator
-    DAQTransferFactory(const DAQTransferFactory&) = delete;
-    DAQTransferFactory& operator=(const DAQTransferFactory&) = delete;
-
-};
+// Implementation of DAQTransferFactory's createChannel method
+std::unique_ptr<IDAQTransfer> DAQTransferFactory::createChannel(const std::string& productType) {
+    if (productType == "A")
+        return std::unique_ptr<IDAQTransfer>(new DAQTransferA());
+    else if (productType == "B")
+        return std::unique_ptr<IDAQTransfer>(new DAQTransferB());
+    else
+        return nullptr;
+}
